+++ src/Models/stock/StockCheck.ts
@@ -1,0 +1,7 @@
+import StockCheckInterface from '../../Interfaces/stock/StockCheck';
+
+export default class StockCheck {
+    constructor(parameters: StockCheckInterface) {
+        Object.assign(this, parameters);
+    }
+}